+++ internal/pkg/httpsrv/handler_websocket.go
@@ -28,9 +28,7 @@ func (s *Server) handlerWebSocket(w http.ResponseWriter, r *http.Request) {
 
 	// Start WS.
 	var upgrader = websocket.Upgrader{
-		CheckOrigin: func(r *http.Request) bool {
-			return true
-		},
+		CheckOrigin: s.checkOrigin,
 	}
 
 	c, err := upgrader.Upgrade(w, r, nil)